Feat: Integrate caching to reduce server load. Especially Agenda-Detail page
Previousl, the agenda-detail page sent large queries to retrieve statical information.
However, as the amount of vote entity increased, there was an issue with long request duration and time out error.
To addressthis, the statistical informations of vote are now recoreded in an LRU cache, and some changes were made to the DTOs for both the server and the client.